refactor(content): import content directly
no reason to create components that accept props and then pass data obtusely through other components. In this project it makes more sense to have the component be responsible for its own dynamic data.

If these components were to be re-used in the same app we would still prop drill